Put CachedOperation in helpers.
